+++ src/content/init-discovery.js
@@ -8,7 +8,7 @@ import settingsPage from '../settings';
  * @returns {Discovery}
  */
 export function initDiscovery(options, data) {
-    const { settings, progressbar, getRaw } = options;
+    const { settings, progressbar, raw } = options;
     const { darkmode = 'auto' } = settings;
     const discovery = new Widget(options.node, null, {
         inspector: true,
@@ -16,20 +16,57 @@ export function initDiscovery(options, data) {
         darkmodePersistent: false,
         styles: [{ type: 'link', href: chrome.runtime.getURL('index.css') }]
     });
+    const copyToClipboard = async function(text) {
+        try {
+            await navigator.clipboard.writeText(text);
+        } catch (err) {
+            console.error(err); // eslint-disable-line no-console
+        }
+
+        discovery.flashMessage('JSON copied to clipboard', 'success');
+    };
+    const downloadAsFile = function(text) {
+        const blob = new Blob([text], { type: 'application/json' });
+        const location = (window.location.hostname + window.location.pathname)
+            .replace(/[^a-z0-9]/gi, '-')
+            .replace(/-$/, '');
+
+        const link = document.body.appendChild(document.createElement('a'));
+        link.download = location.endsWith('-json') ? location.replace(/-json$/, '.json') : location + '.json';
+        link.href = window.URL.createObjectURL(blob);
+        link.click();
+        link.remove();
+    };
+    const renderEl = (config, data, context) => {
+        const fragment = document.createDocumentFragment();
+
+        return discovery.view.render(fragment, config, data, context)
+            .then(() => fragment.firstChild);
+    };
 
     discovery.apply(router);
+    discovery.setPrepare((_, { addQueryHelpers }) => {
+        addQueryHelpers({
+            weight(current, prec = 1) {
+                const unit = ['bytes', 'kB', 'MB', 'GB'];
+
+                while (current > 1000) {
+                    current = current / 1000;
+                    unit.shift();
+                }
+
+                return current.toFixed(prec).replace(/\.0+$/, '') + unit[0];
+            }
+        });
+    });
 
     discovery.flashMessagesContainer = utils.createElement('div', 'flash-messages-container');
     discovery.dom.container.append(discovery.flashMessagesContainer);
     discovery.flashMessage = (text, type) => {
-        const fragment = document.createDocumentFragment();
-
-        discovery.view.render(fragment, {
+        renderEl({
             view: `alert-${type}`,
             content: 'text'
-        }, text).then(() => {
-            const el = fragment.firstChild;
-
+        }, text).then((el) => {
             discovery.flashMessagesContainer.append(el);
             setTimeout(() => el.classList.add('ready-to-remove'), 1250);
             setTimeout(() => el.remove(), 1500);
@@ -45,11 +82,63 @@ export function initDiscovery(options, data) {
         }
     ]);
 
-    discovery.view.define('raw', el => {
-        el.textContent = getRaw();
-    }, { tag: 'pre' });
+    discovery.view.define('raw', (el, _, raw) => {
+        const contentEl = el.appendChild(document.createElement('pre'));
 
-    discovery.page.define('raw', 'raw');
+        contentEl.className = 'content';
+
+        if (raw.firstSlice) {
+            contentEl.append(raw.firstSlice);
+            discovery.view.render(el, {
+                view: 'alert-warning',
+                className: 'too-big-json',
+                content: [
+                    'text:`JSON is too big (${size.weight()} bytes), only first ${firstSlice.size().weight()} is shown. Output the entire JSON may cause to browser freezing for a while. `',
+                    {
+                        view: 'button',
+                        content: 'text:"Show all"',
+                        onClick(el) {
+                            const alertEl = el.parentNode;
+                            alertEl.textContent = 'Output entire JSON...';
+                            setTimeout(() => {
+                                contentEl.append(raw.json.slice(raw.firstSlice.length));
+                                alertEl.remove();
+                            }, 50);
+                        }
+                    }
+                ]
+            }, raw);
+        } else {
+            contentEl.append(raw.json);
+        }
+    });
+
+    discovery.page.define('raw', {
+        view: 'context',
+        data: () => raw,
+        content: [
+            {
+                view: 'page-header',
+                content: [
+                    {
+                        view: 'button',
+                        content: 'text:"Copy to clipboard"',
+                        onClick(_, { json }) {
+                            copyToClipboard(json);
+                        }
+                    },
+                    {
+                        view: 'button',
+                        content: 'text:"Download as file"',
+                        onClick(_, { json }) {
+                            downloadAsFile(json);
+                        }
+                    }
+                ]
+            },
+            'raw'
+        ]
+    });
 
     discovery.nav.append({
         when: () => discovery.pageId !== 'default',
@@ -64,19 +153,6 @@ export function initDiscovery(options, data) {
         content: 'text:"Make report"',
         onClick: () => discovery.setPage('report')
     });
-    discovery.apply(navButtons.inspect);
-    discovery.nav.menu.append({
-        content: 'text:"Download JSON"',
-        onClick: el => {
-            const blob = new Blob([getRaw()], { type: 'application/json' });
-            const location = (window.location.hostname + window.location.pathname)
-                .replace(/[^a-z0-9]/gi, '-')
-                .replace(/-$/, '');
-
-            el.download = location.endsWith('-json') ? location.replace(/-json$/, '.json') : location + '.json';
-            el.href = window.URL.createObjectURL(blob);
-        }
-    });
     discovery.nav.append({
         when: () => discovery.pageId !== 'raw',
         content: 'text:"JSON"',
@@ -85,21 +161,18 @@ export function initDiscovery(options, data) {
             el.title = 'Show JSON as is';
         }
     });
+    discovery.apply(navButtons.inspect);
+    discovery.nav.menu.append({
+        content: 'text:"Download JSON as file"',
+        onClick: (_, { hide }) => hide() & downloadAsFile(raw.json)
+    });
     discovery.nav.menu.append({
         content: 'text:"Copy JSON to clipboard"',
-        onClick: async function() {
-            try {
-                await navigator.clipboard.writeText(getRaw());
-            } catch (err) {
-                console.error(err); // eslint-disable-line no-console
-            }
-
-            discovery.flashMessage('Copied!', 'success');
-        }
+        onClick: (_, { hide }) => hide() & copyToClipboard(raw.json)
     });
     discovery.nav.menu.append({
         content: 'text:"Settings"',
-        onClick: () => discovery.setPage('settings')
+        onClick: (_, { hide }) => hide() & discovery.setPage('settings')
     });
 
     return discovery.setDataProgress(